packages/ai-services/src/vector-similarity/requirements.txt
# backend/vector-search/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-arango==7.5.9
numpy==1.24.3